New invoker, TreeViewInvoker
This new invoker is able to handle all the palettes of one treeview,
removing the need of one invoker per cell renderer (CellRendererInvoker).
This simplifies the code and makes it more efficient.
Also removes the logic from the CellRendererIcon, which
should only care about drawing itself. [1]

This commit also removes the invoker of CellRendererIcon.

[1] https://developer.gnome.org/gtk3/stable/GtkCellRenderer.html