remove GSource used by TimeoutAlert
Remove the timeout when the user clicks one of the buttons.  Otherwise,
the TimeoutAlert object persists, and an OK response occurs when
the timeout reaches zero.

Test timeout with a comparison rather than equality, in case a float is
provided.  Otherwise the timeout never expires.

Immediately deliver an OK response if the timeout provided is less than
one second.