Add a basic, high level abstraction to the collab system
This commit add a high level collab abstraction.  The abstraction
runs over a text channel, so it works with the telepathy changes.

The abstraction automatically sets up new buddies by sending them
the state from the leader.  The abstraction allows activities to
post json encodeable objects to the other users.