High level wrapper over collaberation (for discussion)
This commit introduces the CollabWrapper.  This is a high level
wrapper over the collaberation system.  It uses telepathy text
channels (rather than deprecated tubes used in most activities)
as well as creating file transfer channels to transfer large
messages in a 1-to-1 manner.

For activities, the CollabWrapper provides a simple api.  It syncs
the state when a user joins the activity.  It allows for posting
messages to all other buddies.  It also allows for initiating a
file transfer directly to a buddy, from either a string in memory
or a file on the file system.

TODO:

* Further test resuming a shared activity
* Discuss if file transfer tubes should be considered backwards
  compatiable, as they popup shell notifications without the
  appropriate patch (https://github.com/sugar/pull/621)
* Test sending a file from a file on the file system
* Test accepting a file to the filesystem