Only give activities telepathy tubes if marked as supported
Tubes API has been deprecated since 0.17.25 [1], which was probably
around 2009.  This has since been removed from telepathy, breaking
collaberation.

To maintain support for older activities, if a gsetting is set,
tubes will still be given.  This is useful for OLPC OS, which is
running an older telepathy version that still supports telepathy
tubes.  This therefore requires a patch for the gsettings schema.

When enabled, activities are still given a text channel.  They can
also make their own channels by using the telepathy connection and
calling CreateChannel on the requests interface.