Add abstraction layer for WebKit1/2 apis
Sugar and activities use WebKit to display content in many places.
Unfortunatly, while the world moves to WebKit2, some Sugar devices
can not come to the party.  Namely, the XOs.  Therefore, it may be
advantageous to add an abstraction layer to support both platforms.

This abstraction layer is basically the WebKit2 api.  This is what
we want for the future - so we can remove the abstraction layer
when we drop WebKit1 support.