Build for either python 2 or python 3
`configure` and `autogen.sh` now accept `--with-python2` or
`--with-python3` options, with the default of python 2.

aa8a5e7 introduced building for python 3, but broke build for python 2,
requiring patching `configure.ac` prior to build.

Code was derived from rpmfusion-packager, found using a GitHub search of
`configure.ac` files for both `python2` and `python3`, and is GPLv3.

Tested on Ubuntu 18.04.

Signed-off-by: James Cameron <quozl@laptop.org>