Fix issue with Buddy.props.key being bytes
Buddy.props.key is supposed to be a str, but is stored as a
dbus.ByteArray before being sent to dbus, decoding it after
the call to GetProperties fixes the issue of it being bytes.

Activities like maze that use Buddy.props.key needed to encode
the received key before comparison worked.

Signed-off-by: Chihurumnaya Ibiam <ibiam@sugarlabs.org>